feature(make-restful-api): Create controllers folder
